refactor(tree): Extract leaf schemas into their own module
